Group models together in a single module
