+++ moirae/estimators/online/joint.py
@@ -6,7 +6,8 @@ import numpy as np
 from scipy.linalg import block_diag
 
 from moirae.models.base import InputQuantities, OutputQuantities, GeneralContainer, HealthVariable, CellModel
-from .utils.model import JointCellModelWrapper, convert_vals_model_to_filter
+from .model.cell import convert_vals_model_to_filter
+from .model.cell import JointCellModelWrapper
 from moirae.estimators.online import OnlineEstimator
 from .filters.base import BaseFilter
 from .filters.distributions import MultivariateRandomDistribution, MultivariateGaussian
