--- moirae/estimators/online/utils/__init__.py
@@ -1,1 +1,0 @@
-"""Utility functions used by many types of estimators"""
